+++ src/main/java/org/jeasy/random/randomizers/range/OffsetDateTimeRangeRandomizer.java
@@ -75,10 +75,10 @@ public class OffsetDateTimeRangeRandomizer extends AbstractRangeRandomizer<Offse
     public OffsetDateTime getRandomValue() {
         long minSeconds = min.toEpochSecond();
         long maxSeconds = max.toEpochSecond();
-        long seconds = (long) nextDouble(minSeconds, maxSeconds);
+        long seconds = nextLong(minSeconds, maxSeconds);
         int minNanoSeconds = min.getNano();
         int maxNanoSeconds = max.getNano();
-        long nanoSeconds = (long) nextDouble(minNanoSeconds, maxNanoSeconds);
+        long nanoSeconds = nextInt(minNanoSeconds, maxNanoSeconds);
         return OffsetDateTime.ofInstant(Instant.ofEpochSecond(seconds, nanoSeconds),
                 EasyRandomParameters.DEFAULT_DATES_RANGE.getMin().getZone());
     }
